% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCohorts.R
\name{createCohorts}
\alias{createCohorts}
\title{Create the exposure and outcome cohorts}
\usage{
createCohorts(
  connectionDetails,
  connection = NULL,
  cdmDatabaseSchema,
  oracleTempSchema = NULL,
  cohortDatabaseSchema,
  cohortTable = "cohort",
  cohortIdsToExcludeFromExecution = c(1100, 1101, 1102, 1103, 1104, 1105, 1106, 2001,
    2005, 2006),
  cohortGroups = ScyllaCharacterization::getUserSelectableCohortGroups(),
  minCellCount = 0,
  incremental = TRUE,
  outputFolder,
  incrementalFolder = file.path(outputFolder, "RecordKeeping")
)
}
\description{
Create the exposure and outcome cohorts
}
\details{
This function will create the target, subgroup, and outcome cohorts using inputs
from this study package and using functions from ScyllaCharacterization
}
