% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaAnalysis.R
\name{synthesizeResults}
\alias{synthesizeResults}
\title{Title}
\usage{
synthesizeResults(allDbsFolder, maExportFolder, maxCores)
}
\arguments{
\item{allDbsFolder}{Folder on the local file system containing the results zip files of all data sites.}

\item{maExportFolder}{A local folder where the meta-anlysis results will be written.}

\item{maxCores}{Maximum number of CPU cores to be used when computing the meta-analyses.}
}
\value{
Does not return a value, but creates a new zip file in the \code{maExportFolder} for the meta-analyses.
}
\description{
Title
}
